Feat: Configuracion de los test y  generar reporte con jacoco

+++ domain/model/src/main/java/co/com/pragma/model/loan/LoanRequest.java
@@ -2,7 +2,6 @@ package co.com.pragma.model.loan;
 
 import lombok.*;
 import lombok.NoArgsConstructor;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -11,7 +10,7 @@ import java.time.LocalDateTime;
 @Builder(toBuilder = true)
 public class LoanRequest {
     private String clientDocument;
-    private BigDecimal amount;
+    private Double amount;
     private Integer termMonths;
     private String loanType;
     private RequestStatus status;
