+++ infrastructure/entry-points/reactive-web/src/main/java/co/com/pragma/api/LoanRequestHandler.java
@@ -15,7 +15,6 @@ import org.springframework.web.reactive.function.server.ServerRequest;
 import org.springframework.web.reactive.function.server.ServerResponse;
 import reactor.core.publisher.Mono;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -41,7 +40,7 @@ public class LoanRequestHandler {
 
             LoanRequest requestEntity = LoanRequest.builder()
                     .clientDocument(dto.getClientDocument())
-                    .amount(BigDecimal.valueOf(dto.getAmount()))
+                    .amount(dto.getAmount())
                     .termMonths(dto.getTermMonths())
                     .loanType(dto.getLoanType())
                     .status(dto.getStatus())
